Remove DpadRecyclerViewDelegate
